+++ bundles/mapping/mapwfs2/plugin/WfsLayerPlugin.ol2.js
@@ -335,12 +335,6 @@ Oskari.clazz.define(
                  */
                 MapLayerVisibilityChangedEvent: function (event) {
                     me.mapLayerVisibilityChangedHandler(event);
-                    if (event.getMapLayer().hasFeatureData() && me.getConfig() && me.getConfig().deferSetLocation) {
-                        me.getSandbox().printDebug(
-                            'sending deferred setLocation'
-                        );
-                        me.mapMoveHandler(event.getMapLayer().getId());
-                    }
                 },
 
                 /**
@@ -741,7 +735,10 @@ Oskari.clazz.define(
                     event.getMapLayer().isVisible()
                 );
 
-                if(event.getMapLayer().isVisible()){
+                if (event.getMapLayer().isVisible() && this.getConfig() && this.getConfig().deferSetLocation) {
+                    this.getSandbox().printDebug(
+                        'sending deferred setLocation'
+                    );
                     this.mapMoveHandler(event.getMapLayer().getId());
                 }
             }